+++ src/main/kotlin/cloud/glitchdev/rfu/utils/Party.kt
@@ -17,6 +17,7 @@ import cloud.glitchdev.rfu.events.managers.PartyFinderEvents.registerJoinRequest
 import cloud.glitchdev.rfu.events.managers.PartyEvents
 import cloud.glitchdev.rfu.events.managers.PartyEvents.registerOnPartyChangeEvent
 import cloud.glitchdev.rfu.events.managers.ShutdownEvents.registerShutdownEvent
+import cloud.glitchdev.rfu.feature.other.ignore.IgnoreUtils
 import cloud.glitchdev.rfu.model.party.FishingParty
 import cloud.glitchdev.rfu.utils.command.AbstractCommand
 import cloud.glitchdev.rfu.utils.command.Command
@@ -191,6 +192,7 @@ object Party : RegisteredEvent {
         registerAllowGameEvent("From ($PLAYER_REGEX): \\[RFUPF\\] I would like to join your party!".toExactRegex()) { _, _, matches ->
             val matchGroups = matches?.groupValues ?: return@registerAllowGameEvent true
             val player = matchGroups[1].removeRankTag()
+            if (IgnoreUtils.getIgnoredEntry().contains(player)) return@registerAllowGameEvent false
             promptInvite(player)
             return@registerAllowGameEvent false
         }